capstone/requirements.txt
streamlit
faiss-cpu
sentence-transformers
groq
pymupdf
python-dotenv